refactor spikewrapper
